% Generated by build script
\name{pdb_polymer}
\alias{pdb_polymer}
\title{Get polymer/entity information for a PDB entry}
\usage{
pdb_polymer()
}
\arguments{
\item{pdb_id}{PDB ID (4-character, e.g., "4HHB")}
}
\value{tibble: pdb_id, entity_id, type, description, organism,}
\description{Get polymer/entity information for a PDB entry}
