% Generated by build script
\name{pdb_search}
\alias{pdb_search}
\title{Search PDB structures}
\usage{
pdb_search()
}
\arguments{
\item{query}{Search query string (e.g., "insulin", "CRISPR")}
\item{rows}{Number of results. Default 10.}
}
\value{tibble: pdb_id, score}
\description{Search PDB structures}
